refactor component helpers
