+++ src/main.ts
@@ -1,5 +1,12 @@
 import "./style.css";
-import { Config, FRAuth, FRCallback, FRStep, FRUser } from "@forgerock/javascript-sdk";
+import {
+  Config,
+  FRAuth,
+  FRCallback,
+  FRStep,
+  FRUser,
+} from "@forgerock/javascript-sdk";
+import { mapCallbacksToComponents } from "./components";
 
 Config.set({
   clientId: "aj-public-sdk-client", // e.g. 'ForgeRockSDKClient'
@@ -14,7 +21,49 @@ Config.set({
 });
 
 const panelFormElem = document.getElementById("panel-form");
-const submitButtonElem = document.getElementById("submit-button");
+
+async function handleSubmit(event: Event, step?: FRStep): Promise<void> {
+  event.preventDefault();
+  step?.callbacks.forEach((cb) => setCallbackValue(cb));
+  await nextStep(step);
+  console.log("submitted!", step);
+}
+
+async function handleLogout(): Promise<void> {
+  try {
+    await FRUser.logout();
+    location.reload();
+  } catch (err) {
+    throw new Error(`Failed to logout: ${err}`);
+  }
+}
+
+async function nextStep(previousStep?: FRStep): Promise<void> {
+  if (!panelFormElem) {
+    throw new Error("Failed to find panel");
+  }
+
+  try {
+    const nextStep = await FRAuth.next(previousStep);
+    console.log("nextStep", nextStep);
+    if (nextStep.type === "Step") {
+      nextStep.callbacks.forEach((cb) =>
+        mapCallbacksToComponents(cb, panelFormElem)
+      );
+      panelFormElem?.removeEventListener("submit", handleSubmit);
+      panelFormElem?.addEventListener("submit", (event) =>
+        handleSubmit(event, nextStep)
+      );
+    } else {
+      // Handle login success or failure
+      console.log("todo: handle login success or failure");
+    }
+  } catch (error) {
+    console.error("nextStep error: ", error);
+    panelFormElem.innerHTML = "Error";
+  }
+}
+
 
 function setCallbackValue(cb: FRCallback): void {
   const cbType = cb.getType();
@@ -32,81 +81,13 @@ function setCallbackValue(cb: FRCallback): void {
   }
 }
 
-async function handleSubmit(event: Event, step?: FRStep): Promise<void> {
-  event.preventDefault();
-  step?.callbacks.forEach((cb) => setCallbackValue(cb));
-  await nextStep(step);
-  console.log("submitted!", step);
+document
+  .getElementById("logout-button")
+  ?.addEventListener("click", handleLogout);
+
+// Start the authentication flow
+try {
+  await nextStep();
+} catch (error) {
+  console.error("Error starting authentication flow: ", error);
 }
-
-async function handleLogout(): Promise<void> {
-  try{
-    await FRUser.logout();
-    location.reload();
-} catch (err) {
-    throw new Error(`Failed to logout: ${err}`);
-}
-}
-
-async function nextStep(previousStep?: FRStep): Promise<void> {
-  try {
-    const nextStep = await FRAuth.next(previousStep);
-    console.log("nextStep", nextStep);
-    if (nextStep.type === "Step") {
-      nextStep.callbacks.forEach(mapCallbacksToComponents);
-      panelFormElem?.removeEventListener("submit", handleSubmit);
-      panelFormElem?.addEventListener("submit", (event) =>
-        handleSubmit(event, nextStep)
-      );
-    } else {
-      // Handle login success or failure
-      console.log("todo: handle login success or failure");
-    }
-  } catch (error) {
-    console.error("nextStep error: ", error);
-    if (panelFormElem) {
-      panelFormElem.innerHTML = "Error";
-    }
-  }
-}
-
-function renderTextInputCallback(cb: FRCallback, inputID: string): void {
-  const prompt = cb.getOutputByName("prompt", "");
-  const wrapper = document.createElement("div");
-
-  const label = document.createElement("label");
-  label.setAttribute("for", prompt);
-  label.innerHTML = prompt;
-
-  const input = document.createElement("input");
-  input.setAttribute(
-    "type",
-    cb.getType() === "PasswordCallback" ? "password" : "text"
-  );
-  // input.setAttribute("id", inputID);
-  input.setAttribute("name", inputID);
-  input.setAttribute("required", "true");
-
-  wrapper.appendChild(label);
-  wrapper.appendChild(input);
-
-  panelFormElem?.insertBefore(wrapper, submitButtonElem);
-}
-
-function mapCallbacksToComponents(cb: FRCallback): void {
-  const cbType = cb.getType();
-  switch (cbType) {
-    case "NameCallback":
-    case "PasswordCallback":
-      renderTextInputCallback(
-        cb,
-        cbType === "NameCallback" ? "username" : "password"
-      );
-      break;
-    default:
-      throw new Error(`Unknown callback type: ${cbType}`);
-  }
-}
-
-await nextStep();
-document.getElementById("logout-button")?.addEventListener("click", handleLogout);
